+++ src/meico/app/gui/VerovioGenerator.java
@@ -35,7 +35,7 @@ class VerovioGenerator {
      * This generates HTML/JavaScript code for a Verovio score rendering.
      * @param mei the music to be rendered
      * @param caller
-     * @return
+     * @return HTML code ready for the browser
      */
     public static String generate(String mei, Object caller) {
         return VerovioGenerator.generate(mei, Settings.useLatestVerovio, Settings.oneLineScore, Settings.scoreFont, caller);
@@ -48,7 +48,7 @@ class VerovioGenerator {
      * @param oneLineScore
      * @param font the font used for score rendering
      * @param caller
-     * @return
+     * @return HTML code ready for the browser
      */
     public static String generate(String mei, boolean useLatestVerovio, boolean oneLineScore, String font, Object caller) {
         String html;
@@ -59,10 +59,10 @@ class VerovioGenerator {
             return "<html>Error: Failed to read prototype HTML. <br><br> " + e.toString() + "</html>";      // print the exception message also to the WebView to give an immediate feedback
         }
 
-        if (!useLatestVerovio || !WebBrowser.isNetAvailable("https://www.verovio.org/javascript/develop/verovio-toolkit.js"))                                                      // if the internal Verovio should be used or no internet connection available
+        if (!useLatestVerovio || !WebBrowser.isNetAvailable("https://www.verovio.org/javascript/develop/verovio-toolkit.js"))   // if the internal Verovio should be used or no internet connection available
             html = html.replace("https://www.verovio.org/javascript/develop/verovio-toolkit.js", caller.getClass().getResource("/resources/Verovio/verovio-toolkit.js").toExternalForm());  // replace the online reference in the HTML by the local reference
 
-        html = html.replace("oneLineScore", Boolean.toString(oneLineScore))                                              // the oneLineScore flag is also set
+        html = html.replace("oneLineScore", Boolean.toString(oneLineScore))                                 // the oneLineScore flag is also set
                 .replace("Leipzig", font)
                 .replace("MeiCode", mei.replace("\n", "").replace("\r", "").replace("\"", "\\\""));         // replace the placeholder string "MeiCode" by actual MEI code; that MEI code should be free of linebreaks and quotation marks ("), hence these are replaced before adding it to the html
 
@@ -70,10 +70,11 @@ class VerovioGenerator {
     }
 
     /**
-     *
+     * This method loads prefabricated SVGs into the verovio.html viewer.
      * @param svgs
+     * @param oneLineScore if set true, no page navigation buttons are generated
      * @param caller
-     * @return
+     * @return HTML code ready for the browser
      */
     public static String generate(SvgCollection svgs, boolean oneLineScore, Object caller) {
         String html;
@@ -84,7 +85,7 @@ class VerovioGenerator {
             return "<html>Error: Failed to read prototype HTML. <br><br> " + e.toString() + "</html>";      // print the exception message also to the WebView to give an immediate feedback
         }
 
-        html = html.replace("oneLineScore", Boolean.toString(oneLineScore));                                             // the oneLineScore flag is also set
+        html = html.replace("oneLineScore", Boolean.toString(oneLineScore));                                // the oneLineScore flag is also set
 
         String svgCode = "";
         for (int i=0; i < svgs.size(); ++i) {